+++ api/src/main/java/com/base/util/Encoder.java
@@ -3,7 +3,7 @@ package com.base.util;
 import org.apache.commons.lang3.StringUtils;
 
 public class Encoder {
-    public static String encode(int num) {
+    public static String encode(long num) {
         String encoded = Base62.fromBase10(num);
         return StringUtils.overlay("555", encoded, 3 - encoded.length(), 3);
     }
